hwp_rotation: Move hwp rotation agent into package
