+++ docker-compose.yml
@@ -170,15 +170,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # HWP Rotation Agent
-  # --------------------------------------------------------------------------
-  ocs-hwp-rotation-agent:
-    image: "ocs-hwp-rotation-agent"
-    build: ./agents/hwp_rotation/
-    depends_on:
-      - "socs"
-
-  # --------------------------------------------------------------------------
   # Smurf file emulator agent
   # --------------------------------------------------------------------------
   ocs-smurf-file-emulator-agent:
